regtest_kill_after_timeout: Add bash 4.2 workaround
With bash 4.2, the killer sub-shell fails to kill the sleep job when
killed (not sure why), resulting in stderr being kept open. When
redirecting stderr to a pipe (as is done in our integration tests), this
results in the pipe getting stuck until all the timeouts are reached.
This commit introduces an ugly hack that redirects /dev/null into the
sleep command's stderr, but does not fix the underlying problem that the
sleep process fails to be killed.