Add --warn-only option to `regtest` command
When `--warn-only=<retcode>` is added as argument to a test, if its
command returns with a code of <retcode>, it will be marked
"run(ignored)" instead of just "run". The `regtest_print_summary`
function –and therefore the `run-tests` command– will not return a
non-zero error code when it encounters a test marked as `*(ignored)`,
but will only print a final warning instead.