framework.sh: Fix kill_after_timeout residual `sleep` process
When the second subshell of `regtest_kill_after_timeout` gets killed,
this does not automatically kill the `sleep` process executing within
that subshell. This would not only lead to an unnecessary `sleep`
process getting attached to the init process, but could in some cases
also lead to other parts of the framework getting stuck (due to the
unclosed stderr pipe?). This patch adds a trap to explicitly kill the
`sleep` process on subshell exit.