[doc] Upload enyx-regtest.html and link to it in the README
The README file previously pointed to the manual's source file, which
appeared incomplete on github and gitlab due to unexpanded `include`
directives. It now points to a pre-generated html stored on ipfs.

The Makefile target `upload-doc` can be used to upload the generated
enyx-regtest.html file to ipfs, after which the REAMDE.adoc file can be
manually updated with the new link.