run_suite: Fix printf locale issue on suite timeout
With certain non-english locales, the `printf` command used to print
error information after the timeout would return an error, resulting in
a premature termination of the test suite set.

This is the result of bash's version of printf taking into account the
decimal separator defined through `LC_NUMERIC`. With french locale
settings for instance, `printf "%f"` would accept "4,2" as its second
argument but not "4.2".