Ensure unique session name / log dir
The session name was just the date in the format `%Y-%m-%d-%H:%M:%S`,
which meant it was possible to have colliding session names if several
test sets were launched within the same second. This did happen in
practice for the "metatests" and would have caused them to fail in the
following commit(s).

We now check to see if a log directory with the same name exists, and
append the nanosecond part of the current time (%N) if it does. This is
not perfect: a race condition still exists when test sets are launched
in parallel since the log directory is not created immediately; but I
think it's good enough for now.