Fix regtest_on_exit (resource cleanup)
When `regtest_on_exit` was being called multiple times, only the last
command was taken into account. This happened due a bad usage of
`BASHPID` inside a subshell (`$(...)`). The subshell was unnecessary
anyway and has been replaced by a simple string.

This bug had the particularly obvious effect (if one cared to check)
that the global temporary directory (/tmp/regtest-XXXXX) was not being
removed on exit.