run_suite: Fix handling of early termination (TERM/INT)
Early termination (e.g. through signals TERM or INT) was not properly
handled and resulted in a messy or incomplete summary, or even no
summary at all.

Now, the summary is always printed and contains all suites and tests
having been run up to the time of the interruption. The suite which was
running at the moment the signal was received is marked as "killed".