print_summary: Count suite failures towards return code
There are situations in which a test suite can fail without any of the
underlying test cases failing. In particular, if the test suite is
killed or times out while in between two test cases, the test suite will
report a failure ("killed" or "timeout") even though all of its
evaluated test cases will have succeeded (note: unevaluated test cases
cannot be accounted for).

This commit ensures a suite failure is not only reported in the summary,
but also in the return code.