on_exit_handler: Speed-up killing of children
Turns out the `ps` command is _slow_, probably in part because it must
look at the PPIDs of all processes running on the system. This commit
makes it so we use /proc/<pid>/task/<pid>/children when available, which
is _much_ faster.

This has a very measurable impact on short tests (shorter than 1
second). On my system the time taken by `tests/run-metatests` has been
reduced to 5 seconds, from 8 previously.

Note: We can't just use `jobs -p` in the general case since its output
does not include process substitutions (except to some extent in bash 5
where it includes the last one) or all processes in a pipeline. See:

     16e20cc kill_children_on_exit: Fix bash 5 incompatibility