framework: Add second suite timeout
A second suite timeout can now be requested by setting
'REGTEST_SUITE_TIMEOUT' to a string of the form `[<t1>]+<t2>`. When set,
the first timeout will only have for effect to report an error and the
suite will not be killed. Only if the second timeout gets triggered will
the suite be killed.