checksum-files: With -g|--generate, check existing checksums
When 'regtest_generate' is set (e.g. with -g|--generate) and a reference
file's checksum is already recorded in the checksum file and does not
correspond to the actual checksum of the file —which presumably has just
been generated, comment out (with a '#') the invalid lines in the
checksum file before appending the new checksum to the checksum file and
returning an error.

TODO: Tests!